+++ packages/xarc-app-dev/lib/dev-admin/dev-fastify.js
@@ -2,19 +2,12 @@
 
 /* eslint-disable no-console, no-magic-numbers */
 
-const fastifyPlugin = require("fastify-plugin");
-const archetype = require("@xarc/app/config/archetype");
 const Middleware = require("./middleware");
 const Url = require("url");
 const mime = require("mime");
 const fs = require("fs");
 
 async function register(fastify) {
-  if (!archetype.webpack.devMiddleware) {
-    console.error("dev-hapi plugin was loaded but WEBPACK_DEV_MIDDLEWARE is not true. Skipping.");
-    return;
-  }
-
   const middleware = new Middleware({
     baseUrl: () => {
       return Url.format({
@@ -32,10 +25,11 @@ async function register(fastify) {
     await middleware.process(request.raw, reply, {
       skip: () => {},
       replyHtml: html => {
+        console.log("replying", html);
         reply
           .code(200)
           .header("Content-Type", "text/html")
-          .send(html);
+          .send(`<!DOCTYPE html>${html}`);
       },
       replyNotFound: () => {
         reply.callNotFound();
@@ -44,10 +38,9 @@ async function register(fastify) {
         reply.send(err);
       },
       replyStaticData: data => {
-        const type = mime.lookup(request.url);
+        const type = mime.getType(request.url);
         if (type) {
-          const charset = mime.charsets.lookup(type);
-          reply.header("Content-Type", type + (charset ? `; charset=${charset}` : ""));
+          reply.header("Content-Type", type);
         }
         reply.code(200).send(data);
       },
@@ -59,10 +52,9 @@ async function register(fastify) {
           reply.code(404);
           return;
         }
-        const type = mime.lookup(name);
+        const type = mime.getType(name);
         if (type) {
-          const charset = mime.charsets.lookup(type);
-          reply.header("Content-Type", type + (charset ? `; charset=${charset}` : ""));
+          reply.header("Content-Type", type);
         }
         reply.code(200).send(data);
       }
@@ -72,6 +64,4 @@ async function register(fastify) {
   fastify.use(middleware.hotMiddleware);
 }
 
-module.exports = fastifyPlugin(register, {
-  name: "electrode-dev-fastify"
-});
+module.exports = register;