+++ packages/xarc-webpack/lib/partials/isomorphic.js
@@ -2,12 +2,13 @@
 
 const archetype = require("@xarc/app/config/archetype");
 const IsomorphicLoaderPlugin = require("isomorphic-loader/lib/webpack-plugin");
+const isomorphicConfig = require("isomorphic-loader/lib/config");
 const Path = require("path");
 const { babel } = archetype;
 
 module.exports = function(opts) {
   const target = babel.target !== "default" ? `-${babel.target}` : "";
-  const configFile = Path.resolve(`.isomorphic-loader-config${target}.json`);
+  const configFile = Path.resolve(isomorphicConfig.configFile.replace(".json", `${target}.json`));
   return {
     plugins: [
       new IsomorphicLoaderPlugin({